test(lotto):update test with Describe - Context - It Pattern on LottoManager
- DCI 패턴을 적용해서 테스트 계층화
