+++ src/main/java/lotto/UI/LottoPrinter.java
@@ -12,8 +12,34 @@ public class LottoPrinter {
     public final String WINNING_NUMBERS_PATTERN_ERROR = "[ERROR] 입력 형식에 맞게 숫자를 입력해주세요.";
     public final String WINNING_NUMBER_RANGE_ERROR = "[ERROR] 1~45 범위 내의 숫자를 입력해주세요.";
 
+    public final int PERCENT_CONSTANT = 100;
+    public void printPurchaseAmountMessage() {
+        final String INPUT_AMOUNT_MESSAGE = "구입금액을 입력해 주세요.";
+        System.out.println(INPUT_AMOUNT_MESSAGE);
+    }
     public void printLottoNumbers(List<Integer> lottoNumber) {
         System.out.println(lottoNumber);
+    }
+    public void printPurchaseCompletionMessage(int lottoCount) {
+        final String PURCHASE_MESSAGE = "개를 구매했습니다.";
+        System.out.println(lottoCount + PURCHASE_MESSAGE);
+    }
+
+    public void printGetWinningNumberMessage() {
+        final String INPUT_WINNING_NUMBER_MESSAGE = "당첨 번호를 입력해 주세요.";
+        System.out.println();
+        System.out.println(INPUT_WINNING_NUMBER_MESSAGE);
+    }
+    public void printGetBonusNumberMessage() {
+        final String INPUT_BONUS_NUMBER_MESSAGE = "보너스 번호를 입력해 주세요.";
+        System.out.println();
+        System.out.println(INPUT_BONUS_NUMBER_MESSAGE);
+    }
+
+    public void printStatisticMessage() {
+        System.out.println();
+        System.out.println("당첨 통계");
+        System.out.println("---");
     }
 
     public void printWinningResult(LottoRank lottoRank) {
@@ -29,42 +55,19 @@ public class LottoPrinter {
     }
 
     public void printProfit(double totalProfit) {
-        double totalProfitPercent = totalProfit * 100;
+        double totalProfitPercent = totalProfit * PERCENT_CONSTANT;
         System.out.print("총 수익률은 ");
         System.out.print(String.format("%.1f", totalProfitPercent));
         System.out.print("%입니다.");
     }
 
-    public void printBonusNumberMessage() {
-        final String INPUT_BONUS_NUMBER_MESSAGE = "보너스 번호를 입력해 주세요.";
-        System.out.println();
-        System.out.println(INPUT_BONUS_NUMBER_MESSAGE);
-    }
+
 
     public void printErrorMessage(String errorMessage) {
         System.out.println(errorMessage);
     }
 
-    public void printPurchaseAmountMessage() {
-        final String INPUT_AMOUNT_MESSAGE = "구입금액을 입력해 주세요.";
-        System.out.println(INPUT_AMOUNT_MESSAGE);
-    }
 
-    public void printPurchaseCompletionMessage(int lottoCount) {
-        final String PURCHASE_MESSAGE = "개를 구매했습니다.";
-        System.out.println(lottoCount + PURCHASE_MESSAGE);
-    }
 
-    public void printGetWinningNumberMessage() {
-        final String INPUT_WINNING_NUMBER_MESSAGE = "당첨 번호를 입력해 주세요.";
-        System.out.println();
-        System.out.println(INPUT_WINNING_NUMBER_MESSAGE);
-    }
-
-    public void printStatisticMessage() {
-        System.out.println();
-        System.out.println("당첨 통계");
-        System.out.println("---");
-    }
 
 }
